+++ src/main/java/com/bookripple/api/domain/review/controller/ReviewMemoController.java
@@ -1,11 +1,11 @@
 package com.bookripple.api.domain.review.controller;
 
-import com.bookripple.api.common.code.CommonSuccessCode;
-import com.bookripple.api.common.response.ApiResponse;
 import com.bookripple.api.domain.review.dto.ReviewMemoResDto.MyReviewMemoList;
 import com.bookripple.api.domain.review.service.ReviewMemoService;
+import com.bookripple.api.global.code.CommonSuccessCode;
 import com.bookripple.api.global.dto.GlobalDto.ContentReq;
 import com.bookripple.api.global.dto.GlobalDto.IdRes;
+import com.bookripple.api.global.response.ApiResponse;
 import com.bookripple.api.global.validation.ValidationGroups;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
